Move and rename application
