+++ src/main/java/com/clickbus/restapi/Application.java
@@ -1,12 +1,11 @@
-package com.clickbus.restapi.test;
+package com.clickbus.restapi;
 
 import org.springframework.boot.SpringApplication;
 import org.springframework.boot.autoconfigure.SpringBootApplication;
 
 @SpringBootApplication
-public class RestapiApplication {
-
+public class Application {
     public static void main(String[] args) {
-        SpringApplication.run(RestapiApplication.class, args);
+        SpringApplication.run(Application.class, args);
     }
 }
